+++ CHANGELOG.md
@@ -11,8 +11,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ### Added
 
-- SC-8512 - route to fix event duplications
-
+- SC-8512 - route to fix event duplications & setting IS_MIGRATION in config.js
 ## [25.3.1] - 2020-12-18
 
 ## Added